refactor(claw-server-rust): group HTTP and MCP APIs
