+++ packages/browseros-agent/apps/claw-server-rust/src/api/http/settings.rs
@@ -1,0 +1,34 @@
+use super::error;
+use crate::{AppState, error::CanonicalError, error::RequestId};
+use axum::{
+    Extension, Json,
+    extract::{State, rejection::JsonRejection},
+    http::StatusCode,
+};
+use claw_api::models::{TelemetryState, UpdateTelemetryRequest};
+
+pub(super) async fn telemetry(State(state): State<AppState>) -> Json<TelemetryState> {
+    Json(to_contract_state(state.telemetry.get_state().await))
+}
+
+pub(super) async fn update_telemetry(
+    Extension(request_id): Extension<RequestId>,
+    State(state): State<AppState>,
+    payload: Result<Json<UpdateTelemetryRequest>, JsonRejection>,
+) -> Result<Json<TelemetryState>, CanonicalError> {
+    let Json(payload) = payload.map_err(|_| {
+        error(
+            &request_id,
+            StatusCode::BAD_REQUEST,
+            "invalid_request",
+            "consent must be a boolean",
+        )
+    })?;
+    Ok(Json(to_contract_state(
+        state.telemetry.set_consent(payload.consent).await,
+    )))
+}
+
+fn to_contract_state(state: crate::telemetry::TelemetryState) -> TelemetryState {
+    TelemetryState::new(state.distinct_id, state.enabled, state.consent)
+}
