+++ packages/browseros-agent/apps/claw-server-rust/src/api/mcp/effects/session_naming.rs
@@ -1,4 +1,4 @@
-use crate::mcp::{
+use crate::api::mcp::{
     dispatch::{ToolEffect, ToolEffectContext},
     naming::{build_session_group_title, client_prefix_from_slug},
 };
@@ -38,7 +38,7 @@ mod tests {
     use serde_json::json;
 
     async fn apply_result(
-        call: &crate::mcp::dispatch::ToolCall,
+        call: &crate::api::mcp::dispatch::ToolCall,
         result: &ToolResult,
     ) -> anyhow::Result<ToolResult> {
         Ok(apply(ToolEffectContext {
@@ -60,7 +60,8 @@ mod tests {
     #[tokio::test]
     async fn appends_exactly_five_trailing_tips_and_preserves_image_content() -> anyhow::Result<()>
     {
-        let call = crate::mcp::test_support::tool_call("tabs", json!({ "action": "list" })).await?;
+        let call =
+            crate::api::mcp::test_support::tool_call("tabs", json!({ "action": "list" })).await?;
         let image = ContentBlock::image("aW1hZ2U=".to_string(), "image/png".to_string());
         let original = ToolResult {
             content: vec![ContentBlock::text("ok"), image.clone()],
@@ -89,7 +90,8 @@ mod tests {
 
     #[tokio::test]
     async fn errors_do_not_consume_a_nudge_and_rename_stops_tips() -> anyhow::Result<()> {
-        let call = crate::mcp::test_support::tool_call("tabs", json!({ "action": "new" })).await?;
+        let call =
+            crate::api::mcp::test_support::tool_call("tabs", json!({ "action": "new" })).await?;
         let error = ToolResult::error("failed");
         assert_result_eq(&apply_result(&call, &error).await?, &error);
 
@@ -108,9 +110,9 @@ mod tests {
     #[tokio::test]
     async fn separate_sessions_have_independent_nudge_budgets() -> anyhow::Result<()> {
         let first =
-            crate::mcp::test_support::tool_call("tabs", json!({ "action": "list" })).await?;
+            crate::api::mcp::test_support::tool_call("tabs", json!({ "action": "list" })).await?;
         let second =
-            crate::mcp::test_support::tool_call("tabs", json!({ "action": "list" })).await?;
+            crate::api::mcp::test_support::tool_call("tabs", json!({ "action": "list" })).await?;
         let success = ToolResult::text("ok", None);
 
         for _ in 0..5 {
